@@ -4,7 +4,7 @@ description: Demonstrates how to use the device_info_plus plugin.
 dependencies:
   flutter:
     sdk: flutter
-  device_info_plus: ^9.0.0
+  device_info_plus: ^9.0.2
 
 dev_dependencies:
   flutter_driver:
@@ -19,4 +19,4 @@ flutter:
   uses-material-design: true
 
 environment:
-  sdk: '>=2.18.0 <4.0.0'
+  sdk: '>=3.0.0 <4.0.0'
